backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary==2.9.10",
    "datasets>=4.0.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.116.1",
    "mistralai>=1.9.3",
    "pydantic>=2.11.7",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
    "matplotlib>=3.10.5",
    "scikit-learn>=1.7.1",
    "pgvector>=0.4.1",
    "seaborn>=0.13.2",
]
